Migrations updates

+++ users/models.py
@@ -34,7 +34,7 @@ class User(AbstractUser):
 
 
 class Troop(models.Model):
-    troop_number = models.PositiveSmallIntegerField(unique=True)
+    troop_number = models.IntegerField(unique=True)
     troop_name = models.CharField(max_length=300, blank=True)
 
     troop_cookie_coordinator = models.CharField(max_length=300, null=True)
